use std.log & print progress on Release
Since Zig has a built in logging system with levels that depend on the build option, I've switched on using those.
Also, if we are on Release which doesn't show any logs except errors, print a pretty progress that updates every tick, to give the user feedback on the download.